more role specific actions

+++ components/purchaserequest/actions/delete_request.tsx
@@ -1,14 +1,22 @@
 "use server";
 
+import { getUserSession } from "@/lib/auth";
+
 //ansyc function to delete a purchase request
 export async function deletePR({ id }: { id: string }) {
-    const req = await fetch(
-        `https://shim-ventare.vercel.app/api/purchaserequests/${id}`,
-        {
-          method: "DELETE",
-          headers: { "Content-Type": "application/json" },
-        }
-    );
-    const res = await req.json();
-    return res;
+  //get user 
+  const user = await getUserSession();
+  //if user is not an admin, return error
+  if(user.role !== "Admin") {
+    return { error: "You must be an admin to delete a purchase request" };
+  }
+  const req = await fetch(
+      `https://shim-ventare.vercel.app/api/purchaserequests/${id}`,
+      {
+        method: "DELETE",
+        headers: { "Content-Type": "application/json" },
+      }
+  );
+  const res = await req.json();
+  return res;
 }
